make util for available findings
